+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.0.25
+
+* exposing storage service.
+
+## 0.0.24
+
+* fix media uploading..
+
+## 0.0.23
+
+* fix media uploading.
+
+
 ## 0.0.22
 
 * fix `DisplayMedia`.